Split out first test
